refactor: mutation 활용
