+++ src/domain/api/index.ts
@@ -1,11 +1,18 @@
-import kyApi from "@/src/common/apis/ky-api"
+import KyInstance from "@/src/common/apis/ky-instance"
 import { IUserData } from "@/src/domain/types/user"
 
-const signupApi = {
+interface DummyResponse {
+  test: string
+}
+
+const signApi = {
   signup: async (userData: IUserData) => {
-    const response = await kyApi("post", "/api/auth/profile", userData)
+    const response = await KyInstance.post<DummyResponse>("/api/auth/profile", {
+      json: userData,
+    }).json()
+
     return response
   },
 }
 
-export default signupApi
+export default signApi
