+++ src/domain/oauth/social-login.tsx
@@ -1,20 +1,11 @@
 "use client"
 
+import { useMutation } from "@tanstack/react-query"
 import { setCookie } from "cookies-next"
-import ky from "ky"
 import { useRouter } from "next/navigation"
 import { useEffect } from "react"
+import { socialLogin } from "@/src/domain/signup/apis"
 import useSignup from "@/src/domain/signup/context/signup-context"
-
-interface IAuthResponse {
-  accessToken: string
-  isSignup: boolean
-  name: string
-  phoneNumber: string
-  birthday: string
-  gender: string
-  email: string
-}
 
 interface ISocialType {
   socialType: string
@@ -31,72 +22,69 @@ const SocialLogin = ({ socialType }: ISocialType) => {
 
   const redirectUri = redirectUriMap[socialType]
 
+  const { mutate, isPending, isError } = useMutation({
+    mutationFn: async (code: string) => socialLogin(socialType, code, redirectUri),
+    onSuccess: data => {
+      const formattedBirthday = data?.birthday?.split("T")[0]
+      const [year, month, day] = formattedBirthday ? formattedBirthday.split("-") : ["", "", ""]
+
+      const userData = {
+        name: data.name ?? "",
+        phoneNumber: data.phoneNumber ?? "",
+        birthday: formattedBirthday ?? "",
+        gender: data.gender ?? "",
+        email: data.email ?? "",
+        accessToken: data.accessToken ?? "",
+      }
+
+      if (data.isSignup === true) {
+        setCookie("accessToken", data.accessToken, {
+          path: "/",
+          httpOnly: false,
+          maxAge: 60 * 60 * 24 * 7,
+        })
+        router.push("/")
+      } else if (data.isSignup === false) {
+        setUser(userData)
+        handleYear(year)
+        handleMonth(month)
+        handleDay(day)
+
+        const pages = ["name", "phone", "birthday", "gender", "check"]
+        const keyMapping = { name: "name", phone: "phoneNumber", birthday: "birthday", gender: "gender" }
+
+        const updatedFilterPages = pages.filter(page => {
+          const mappedKey = keyMapping[page as keyof typeof keyMapping]
+          return mappedKey && !(userData as any)[mappedKey]
+        })
+
+        if (!updatedFilterPages.includes("check")) {
+          updatedFilterPages.push("check")
+        }
+
+        // console.log("🚀 [필터 페이지 최종]:", updatedFilterPages)
+        setFilterPages(updatedFilterPages)
+
+        router.push(`/${updatedFilterPages[0]}`)
+      }
+    },
+  })
+
   useEffect(() => {
     const code = new URL(window.location.href).searchParams.get("code")
-    const baseUrl = process.env.NEXT_PUBLIC_BASE_API_URL
 
-    const fetchAuthData = async () => {
-      if (code) {
-        await ky
-          .post(`${baseUrl}/auth/social/${socialType}`, {
-            json: { code, redirectUri },
-            headers: { Accept: "application/json" },
-          })
-
-          .json<IAuthResponse>()
-          .then(data => {
-            setCookie("accessToken", data.accessToken, {
-              path: "/",
-              httpOnly: false,
-              maxAge: 60 * 60 * 24 * 7,
-            })
-            const formattedBirthday = data?.birthday?.split("T")[0]
-            const [year, month, day] = formattedBirthday ? formattedBirthday.split("-") : ["", "", ""]
-
-            const userData = {
-              name: data.name ?? "",
-              phoneNumber: data.phoneNumber ?? "",
-              birthday: formattedBirthday ?? "",
-              gender: data.gender ?? "",
-              email: data.email ?? "",
-            }
-
-            // console.log("userData", userData)
-
-            if (data.isSignup === true) {
-              router.push("/")
-            } else if (data.isSignup === false) {
-              setUser(userData)
-              handleYear(year)
-              handleMonth(month)
-              handleDay(day)
-
-              const pages = ["name", "phone", "birthday", "gender", "check"]
-              const keyMapping = { name: "name", phone: "phoneNumber", birthday: "birthday", gender: "gender" }
-
-              const updatedFilterPages = pages.filter(page => {
-                const mappedKey = keyMapping[page as keyof typeof keyMapping]
-                return mappedKey && !(userData as any)[mappedKey]
-              })
-
-              if (!updatedFilterPages.includes("check")) {
-                updatedFilterPages.push("check")
-              }
-
-              // console.log("🚀 [필터 페이지 최종]:", updatedFilterPages)
-              setFilterPages(updatedFilterPages)
-
-              router.push(`/${updatedFilterPages[0]}`)
-            }
-          })
-      }
+    if (code) {
+      mutate(code)
     }
-
-    fetchAuthData()
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
-  return <div>로그인 중입니다...</div>
+  if (isPending) {
+    return <div>로그인 중입니다...</div>
+  }
+  if (isError) {
+    return <div>로그인 실패...</div>
+  }
 }
 
 export default SocialLogin
